pyproject.toml
[project]
name = "neutral-detection-agents-workflow"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "fastembed>=0.5.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "tavily-python>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["src"]}}
